fix(model): balance_history.balanceChange DECIMAL(12,2) to match balances
history_change was INTEGER while user_balance is DECIMAL(12,2); fractional sell
credits (e.g. 12.53) were rounded when logged, so the audit trail diverged from
the real balance change. Use DECIMAL(12,2).

+++ backend/src/models/balanceHistory.js
@@ -20,7 +20,7 @@ module.exports = sequelize.define(
         },
         history_change: {
             field: 'balanceChange',
-            type: Sequelize.INTEGER,
+            type: Sequelize.DECIMAL(12, 2),
         },
         history_extraData: {
             field: 'extraData',
